requirements.txt
pywebview
Pillow